requirements.txt
playwright
beautifulsoup4
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pandas
openpyxl
python-dotenv
schedule 
